INTLY-1362 Allow managedServices to be specified in JSON config
A change is being added to the web app to allow for a
managedServices configuration, this will be used to allow users to
specify which managed services in Integreatly they require for the
walkthrough.

This change updates the schema to allow for this new object along
with an example walkthrough.